requirements.txt
-e .
pytest
aws_cdk.aws_ecs
aws_cdk.aws_ecs_patterns
